fix(update): align shared types and repair test ownership
