+++ scripts/lib/vitest-build-prerequisites.mts
@@ -43,6 +43,7 @@ const runtimeConsumers = [
     "src/cli/update-cli/update-command-migrated.test.ts",
     "src/cli/update-cli/update-command-rollback.test.ts",
     "src/cli/update-cli/update-command-post-update-recovery.test.ts",
+    "src/cli/update-cli/update-command-post-update-repair.test.ts",
     "src/cli/update-cli/update-command-service.integration.test.ts",
   ].map((file) => ({
     file,
