+++ test/vitest/vitest.cli-process-paths.mjs
@@ -21,6 +21,7 @@ export const cliProcessTestFiles = [
   "src/cli/update-cli/update-command-migrated.test.ts",
   "src/cli/update-cli/update-command-rollback.test.ts",
   "src/cli/update-cli/update-command-post-update-recovery.test.ts",
+  "src/cli/update-cli/update-command-post-update-repair.test.ts",
   "src/cli/update-cli/update-command-service.integration.test.ts",
   "src/cli/one-shot-exit.test.ts",
   "src/cli/update-finalization-output.process.test.ts",
